@@ -24,23 +24,26 @@
   "author": "Jason Kadrmas",
   "license": "MIT",
   "dependencies": {
-    "classnames": "^2.1.3",
-    "fbemitter": "^2.0.0",
-    "jquery": "^3.2.1",
-    "moment": "^2.13.0",
-    "react-anything-sortable": "^0.2.2",
-    "react-bootstrap-native-slider": "2.0.1",
-    "react-datepicker": "^0.27.0",
-    "react-select": "^0.5.5",
-    "react-signature-pad": "0.0.6",
-    "react-sortable-items": "0.0.7",
+    "classnames": "^2.2.6",
+    "create-react-class": "^15.6.3",
+    "fbemitter": "^2.1.1",
+    "jquery": "^3.3.1",
+    "moment": "^2.22.0",
+    "prop-types": "^15.6.2",
+    "react-anything-sortable": "^1.7.4",
+    "react-bootstrap-slider": "^2.1.5",
+    "react-datepicker": "^1.6.0",
+    "react-select": "^2.0.0",
+    "react-signature-pad": "^0.0.6",
+    "react-sortable-items": "^0.0.12",
     "react-star-rating": "git+https://github.com/blackjk3/react-star-rating.git",
-    "react-textarea-autosize": "^2.3.1",
+    "react-textarea-autosize": "^7.0.4",
+    "react-draft-wysiwyg": "^1.12.13",
     "reflux": "^0.2.8",
-    "xss": "^0.2.13"
+    "xss": "^1.0.3"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0"
+    "react": "^16.0.0"
   },
   "devDependencies": {
     "babel": "6.23.0",
@@ -51,19 +54,17 @@
     "babel-register": "6.24.1",
     "canvas": "1.6.5",
     "css-loader": "^0.16.0",
-    "draft-js": "^0.10.0",
-    "draftjs-to-html": "^0.6.1",
-    "immutable": "^3.8.1",
+    "draft-js": "^0.10.5",
+    "draftjs-to-html": "^0.8.4",
+    "immutable": "^3.8.2",
     "jsdom": "11.12.0",
-    "mocha": "^3.3.0",
-    "mocha-jsdom": "^1.1.0",
+    "mocha": "^5.2.0",
+    "mocha-jsdom": "^2.0.0",
     "node-libs-browser": "^1.0.0",
     "node-sass": "^4.9.3",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "react-draft-wysiwyg": "1.7.0",
-    "react-hot-loader": "^1.2.7",
-    "react-tools": "0.13.3",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-hot-loader": "^4.3.4",
     "sass-loader": "^2.0.1",
     "source-map": "^0.5.6",
     "style-loader": "^0.12.3",
